+++ src/app/(dashboard)/dashboard/token-saver/page.js
@@ -1,0 +1,5 @@
+import TokenSaverClient from "./TokenSaverClient";
+
+export default function TokenSaverPage() {
+  return <TokenSaverClient />;
+}
